Added Spike Tiles and tweaks
Added spike tiles to the blocks tile palette. It's a game object tile that will damage the player when they come in contact with it, no extra tilemaps required!

Additionally, I added soft collision to all pots and enemies, and made all them prefabs too. Now you can simply update a prefab to apply a change to all instances of it, instead of finding all the times we copy-pasted the Bad Guy to give him the soft collider.

I also investigated with making the breakable pot a tile, but I ran into issues with the animations, and it can only be a 1x1 tile. I decided that the original pots are fine and there's no reason to fix something that ain't broken!

Lastly, I addressed the issue I made with the health bar in the TestDungeon_Room1 scene.