Added Pause Menu, Souls/Godsouls UI, and slight refactoring
You can now press escape to pause the game, which will bring up a UI and pause all game logic. There's also  a UI that shows how many souls/godsouls the player has, and the information is stored in the data manager. However, the numbers shown on screen are not yet connected to the numbers internally. I also refactored how the data manager and healthbar communicate with the player. These systems now subscribe the the onPlayerHealthChange event to trigger their logic, rather than doing so every frame via Update().