Fully Implemented Souls/Godsouls and refactored the entity controllers
Killing enemies will now give the player souls, which are displayed in the UI I added yesterday. This data is remembered across scenes and deaths. I also refactored the melee and ranged entity controllers to inherit from a base entity controller, because of how much the two scripts had in common, which future scripts will probably need to share anyway. This makes the difference in functionality between the two more obvious as well.