Rock Ability Set Complete!
Rock utility and passive abilities have been added, completing the set. The rock utility ability creates a small platform beneath you for 5 seconds, which is helpful for parkour! The rock passive ability gives the player a flat 10 damage resistance, which is subtracted from any damage the player takes (except fire damage). The player will not take damage if the amount of damage is less than the damage resistance stat.