Warp Obelisks
Warp Obelisks are a new GameObject tile that when hit with the player's melee attack, will set the player's spawnpoint. If the player does not have a respawn point set already, it'll default to where the player is placed by default in the scene the game opens in.

Interacting with these obelisks by pressing E will bring you to the Skyhub, but right now only sends a debug message.