Added the Skyhub, long-press interacting, and music
You can now use the Warp Obelisks to warp the the Skyhub! The Skyhub is a place in the clouds where you can talk to the gods who you've lifted the curse of to get their abilities. This is a temporary version of the Skyhub of course. We need to discuss in detail how we actually want it to look.

I changed the behavior of warp obelisks so that you press E to set your spawn, and hold E to warp to the skyhub. When holding E to interact, it'll show a progress bar at the bottom (which was a surprising pain in the ass to make work correctly).

Lastly, I added the music @lulumakesmusic made into the game.