Ability Inventory Part 5 - Finished!
All the textboxes in the details menu can now be set through the ability info, as well as their offense, defense, utility, and passive sprites. Additionally, I added the first original art to the game! The offense, defense, utility, and passive sprites made by Bella are now in the game. For some reason something's wrong with the quality though, which I'll try to fix when I get a chance.