Started converting to XML documentation for Doxygen
Originally, I was going to just write up documentation manually in google docs (as well as improving the comments) but I've instead decided to use a much more efficient method: auto-generating documentation with XML and Doxygen.

Today I converted the scripts on the top level directory and got started on the ability scripts.