No changes. Meta files created
When Unity runs, it creates a .meta file for each new file. Since I've been busy with documentation, I haven't actually opened Unity, so it created a lot of new .meta files. Unfortunately, there appears to be a bug where .meta files are created even outside the assets folder (not supposed to happen), so the Doxygen assets are now full with useless .meta files that are practically empty.