Proper functionality for passive abilities
The proper functionality for passive abilities has finally been introduced. The player no longer has to press 4 to manually activate their passive ability, with it instead activating and deactivating on its own.

Note: the disabling of one of the example ability sets (I think the one with the abilityName "Isis") is bugged and completely breaks the player's health and movement when put into a slot that is not the passive slot.