Mummy and Golem entities
Two new entities. the mummy will appear throughout beginning levels and not sure what to do with the Golem yet. but the Golem is slow but hits hard, knocks back quite a distance, and is hard to knock back itself.  the mummy is slow and does not much damage but is usually in groups.