Fixed Menu/Warping Movement Bug
The player was able to start movement actions even when the player was warping or the game was paused. This was fixed by adding conditions everywhere that was previously causing problems in the player movement script, by checking if the game is paused or if the player is warping. This is probably not be the best way to fix this, especially if more conditions or player actions are added, but it works for now.