Earthquake Projectile and big projectiles refactor
The main part of Geb's offense ability is now complete! Horus can throw a boulder that spawns an earthquake. This earthquake will damage non-player entities within a wide but very short rectangle. I still need to add some visual effects so it's clear what's happening, as well as add the damage increase/slower swing speed that the offense ability is also supposed to provide, but this ability set is almost done!

A lot of changes have been made to the underlying functionality of projectiles as well. Here's a breakdown:
- New RudimentaryMovement scripts.
   - RudimentaryMovement is a script that represents a basic format for simple movement scripts that other scripts can inherit from
   - MoveInDirection is a RudimentaryMovement script that moves an object is a given 2D direction at a consistent rate continuously.
   - InitialThrust is a RudimentaryMovement script that applies a one time thrust in a given 2D direction right as the object is created.
- BasicProjectile has been renamed to BaseProjectile and no longer handles the projectile's movement.
   - Instead, support has been added for new RudimentaryMovement scripts.
   - New private UpdateFacing() function now handles correcting the direction the RudimentaryMovement is moving in, and flipping the sprite.
   - UpdateMethods() has been removed because it's no longer necessary.
   - New destroyOnImpact bool variable, which determines if the projectile destroys itself when it comes in contact with something.
   - New refDamage variable that allows for damage to be referenced by other scripts without editing it, and still showing damage in the Unity Editor.