Mummy walking animation added
The mummy's walking animation has been added! There were some issues implementing it due to missing frames, so the version in the gif send in art a while ago is the one being used (don't worry, it's still good quality, just a slightly different color).

I also fixed a bug with the ground detector where if there is no wall detector attached to the entity, it would throw an error. Now if there's no wall detector, the game assumes the entity is not on a wall. Hopefully this didn't create any additional bugs.