Random Entity Stuff
- I multiplied gravity by 5 in the Physics2D settings, rather than making everything have a gravity scale of 5. Now, most entities have a gravity scale of 1 (the default) rather than 5. Some projectiles now have a gravity scale of 1/5th so that they remain unaffected by this change.
- BaseEntityController.cs no longer breaks when a groundDetector isn't attached, and I cleaned up my code from last time I touched this file.

By assuming that the enemy is never on a wall when there's no wall detector attached, that technically allows an enemy to perform the wall jump bug, but the chance of an enemy pulling that off is extremely unlikely.
By assuming the enemy is always on a wall when there's no wall detector attached, that reintroduces the 0.025 second delay before they can jump (or sometimes move) after touching the ground, but at least it makes the wall jump bug impossible. The delay probably wouldn't be noticeable either.
I decided to go with the latter, but feel free to change it back. The difference is so small either way.