Game Easier to Build
Three scripts were preventing the game from being built because they relied on something that only existed in the unity editor. I added the lines #if UNITY_EDITOR and #endif around everything causing the problems, so that it still works in the editor without stopping the game from being built. It is probably a good idea to have #if UNITY_EDITOR and #endif around everything that relies on UnityEditor, even if it doesn't cause problems, but it should usually be handled automatically, or something IDK I had to look it up. Anything in the future that is only supposed to work in the editor and stops the game from being built should be put between #if UNITY_EDITOR and #endif