Main Menu Added
Main menu & placeholder graphics added!
The main menu is on the top of the build settings list, so when the game is built, it should be the opening scene.
The play button doesn't do anything yet, but it is attached to a script. Save files need to be figured out too.
Add functionality to the main menu whenever you want! This main menu is just a base for any additional features.

I don't know why it's adding the Assets\TextMesh Pro\Resources\Fonts & Materials\LiberationSans SDF - Fallback.asset as a part of this change, I think it keeps doing that automatically.