WallDetector Side Detection Added
The WallDetector now has the option to detect which side of the entity is touching the wall, either the front side or the back side. Before, it couldn't distinguish between the sides. This feature is optional, and I have only added it to BadGuy Patrol just to test it out.
I moved GroundDetector.cs and the wall detector scripts to the entity folder because they are more generic now. IDK if this is right so go ahead and move it back. Also walldetector.cs has been split into 2 different scripts now, and I might even give it a third.
At first, I tried making the wall detector distinguish between the left side and the right side, but when the enemy turned around, it also flipped the wall detector. So now, it is just front and back, and if the difference between left and right is important, that can be determined in the movement scripts.