Sarcophagus Message Animation
I gave the speech bubble for the sarcophagus object that I added yesterday an animation. Now I (slightly) know how to use an animator!
The animator is not perfect: it doesn't apply the alpha fade effect to all the child objects, that has to be done manually. But, even if you don't apply the alpha effect, it's not noticeable.