Particles for Geb's Earthquake
Added particle effect to Geb's earthquake attack. This means that the earthquake damage zone is no longer invisible. The particles should be adjusted at some point to be less random and more spread out, so that way the edges of the damage zone are more defined.
The particle effect is the same as the BreakParticles because I copied it and I don't know how to modify it the way I want.