Changed Geb Bossroom Bounds
Changed the way the bounds of Geb's bossroom work. Instead of having an invisible collider, entities are now stopped from leaving the bossroom with code. This allows for projectiles to continue to travel past where the player/Geb can go. There is a new barrier in the bossroom that is placed out of bounds so projectiles won't fly off forever. This change should hopefully make what can and cannot leave the bounds more consistent.