Added footstep sounds and rebuild FMod project
Footstep sounds now play when the Horus's foot touches the ground in the walking animation (assuming he's on the ground and a sound has been assigned to the tilemap collider using the new TilemapFootstepSFX script). There are footstep sounds for sand, clouds, and generic hard surfaces.

There were some issues with FMod while setting this up, so I had to rebuild the entire project. Fortunately, it should all be working correctly now.