Documentation update and FMOD explanation page
Updated several documentation pages, including the prefabs and scenes pages, as well as some newer scripts that did not receive documentation. Some scripts in the UI folder are still undocumented though.

I also wrote up a detail guide on how to use FMOD. Because the setup and learning can be intimidating, I wanted to make sure an easy to read guide was available.