Fixed ClubFair_2 Audio Bug
Found what was causing the bullet sound whenever ClubFair_2 was entered. An enemy was activating the arrow trap, but the enemy was also resetting the arrow trap timer, so the arrow trap only activated once when the scene was loaded.
I made two changes:
- The arrow trap timer (more generally, SpawnOnTrigger.cs) only gets reset when a projectile is fired, rather than every time an object is detected.
- The arrow trap has a new setting that, when enabled, only the player is detected. This setting is enabled by default.
The arrow trap still only activates when an object enters the trigger zone, so an object must leave and re-enter the trigger zone in order to re-activate the trap (nothing changed).