Data saving to file and file reorganization
Data saving:
When loading the game from the Main Menu scene, the DataManager, with the help of the new FileReadWrite and SerializedData classes, now saves various data such as the player's health, respawn point, soul count, time of day, and volume settings to a file, and will restore those values when the game is launched again through the Main Menu. The player now spawns at their last respawn point when they open the game.

To be clear, this is designed to only work if the game is launched through the Main Menu scene, allowing testing in other scenes to not interfere with the saved file.

File reorganization:
I moved the DataManager script, along with the new FileReadWrite and SerializedData scripts, to the new Data Management folder. I also created a bunch of subfolders under Entities and moved files around in there. I even moved the last scripts sitting in the scripts folder (like ObjectHealth and Soft Collider) into subfolders in Entities.

Other changes:
- Fixed a bug where the player would sometimes spawn in an unexpected location in the Anubis scene outside of the void area, and would fall in a blank world forever.
- Main Menu now loads the player's last respawn point, or if no save file is found, loads the club fair level by default.
- The pause menu now uses Application.Quit() instead of loading the main menu because that a) that's the intended behavior and b) it was messing with the new data saving system.
- TimeOfDay enum in the data manager script is now above, not inside, the DataManager class, allowing it to be used anywhere. This simplifies code in some areas, but is also just the way it should've been this whole time (oops).