Memphis Opening Scene
Added the opening scene for the game. It takes place in a broken city (Memphis). It has some setup where the player learns the controls and learns about Geb. There are no dungeons yet, but there is a dungeon entrance that will be used for the majority of the Memphis. I did not change which level the main menu loads by default.