Memphis Desert Pt. 1
Added the next scene for Memphis. It is not complete yet. It introduces combat. It takes place in the open desert (it could theoretically also take place inside of the city, but I haven't placed any ruined buildings yet).

I changed up the last part of the opening scene when I remembered that daytime/nighttime mechanics are supposed to be important, so the dungeons are not going to come until later. For the dungeons, I'm thinking it will be one large building with a lot of windows (allows for day/night). What will the building be like? I am not sure yet.