Groups for Destructible Walls
The destructible walls now have the option to break as a group rather than only one wall breaking at a time. For example, if you hit one destructible wall tile, all neighboring tiles, and all neighbors of the neighbors, etc. all get destroyed at once. If you want to use this feature, make sure that the walls are siblings (under the same parent) and that destroyGroup is enabled.