Started next scene (indoor level)
I polished a few things in the Memphis scenes so far, and I added a new scene (MemphisIndoors_1). I will expand on it soon. It takes place inside of a big building where there will probably end up being puzzles.
In anticipation of the jump height decreasing, I decreased the player's jump height in this scene, which also allowed me to zoom the camera back in a little bit.