Made Puzzle for First Level
I made a puzzle for the first level. You can try it out in scene MemphisIndoors_1, or start from MemphisDesert_1 and try to make it to Geb. I don't think it's too difficult, but IDK because I'm the one who made it. It's possible to skip some of the puzzle with an ability, but that won't matter.