Added particles to EarthquakeZone
EarthquakeZone (spawned by Geb's offense ability) now spawns the same particles that Geb's earthquakes spawn during his boss battle. The size of the spawn area and the duration of the particle spawning is configured in the script to match the parameters of the EarthquakeZone.

I also re-fixed the bug where the player spawns outside the playable area in the Anubis level (the script was missing from the player, probably because we forgot to restore the it's ID in the cs.meta file).