Added Horus' falling animation & improved animation transitions
Added Horus' falling animation. When Horus' height in the world is decreasing and he's not on the ground, a falling animation will play.

In the process, I've added general improvements to how transitions between animations are done to make them smoother, such the jump, double jump, gliding, and falling animations immediately cancelling when Horus touches the ground, and the running animation cancelling as soon has Horus stops moving.