Improved continuity on MemphisIndoors_2 scene load
When the player entered MemphisIndoors_2 from the exit (either by playing the level backwards or by turning around after entering the next scene), the sun shrine would be in an awkward/impossible spot. Now, when the player enters the scene from the back entrance, the sun shrine starts from where it usually ends, which fixes the continuity of the scene.