End sequence complete
The level is complete! There is still plenty of polishing that needs to get done, but the player can now play through the whole game twice: once without abilities and once with abilities unlocked.

When the player leaves the Skyhub for the first time, they get brought back to the beginning of the level with abilities unlocked, and the player can return to the Skyhub at any time.
I also put a short tutorial message that appears in the Skyhub inventory when you open it for the first time.