Warp obelisk bugs fixed
I changed some of the warp obelisk behavior, and I *think* that *most* of the warp obelisk bugs are fixed. Not sure though. Some of the bugs are fixed though!

There is still that bug with the abilities not unlocking, but I still don't know why that's happening. It only happens some of the time, not always. I'll have to investigate further.