Final commit before first release!
This is the final commit before the first release of Eye of the Nile!
- The game now saves whether you unlocked your abilities or not
- Darkened the background of the title screen
- Fixed a bug where Horus would be stuck in his jump animation, even though on the ground
- Fixed a bug where the abilities the player has equipped would sometimes be in their wrong form (ability in offense slot would be in its passive form, etc)

There is still potentially an issue where the game may not properly remember the ability hotbar layout every time, but I don't know how to replicate it, and if that happens, the abilities should hopefully still work because of the bugfix above. However, there's a chance that if that bug occurs, the player can't use their abilities at all.